+++ CHANGELOG.md
@@ -23,3 +23,4 @@
 - Fix AvatarName if name === '""' or length is empty return address ([\#95](https://github.com/forbole/big-dipper-2.0-solana/issues/95))
 - Add tx details logs component
 - Add tx details instructions ([\#105](https://github.com/forbole/big-dipper-2.0-solana/issues/105))
+- Add volume on token page ([\#112](https://github.com/forbole/big-dipper-2.0-solana/issues/112))